core/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
cryptography==42.0.0
httpx==0.26.0
python-multipart==0.0.6
pyyaml==6.0.1
